n_overlapping is usize

+++ src/indel.rs
@@ -37,14 +37,13 @@
     }
 
     impl IndelRider {
-        pub fn new(snps_buffer: & VecDeque<mutations::Mutation>, n_overlapping: u32, n_samples: usize) -> IndelRider {
+        pub fn new(snps_buffer: & VecDeque<mutations::Mutation>, n_overlapping: usize, n_samples: usize) -> IndelRider {
             let mut groups : Vec<usize> =  vec![0; n_samples*2];
             IndelRider::count_groups(snps_buffer, n_overlapping, & mut groups, n_samples);
             let mut rev_groups : HashMap<usize, Vec<usize>> = HashMap::new();
             //println!("cgroups {:?}", groups);
             // groups has chr samples as indexes and group ids as elements, we need to invert our perspective to iterate easily on groups.
             for (sample, group) in groups.iter().enumerate() {
-                //rev_groups[*group as usize].push(sample as u32); // Mh, use all usize and stop? XXX
                 if rev_groups.contains_key(group) {
                     let mut samples = rev_groups.get_mut(group).unwrap();
                     samples.push(sample);
@@ -55,7 +54,6 @@
             //println!("cgroups {:?}", rev_groups);
             let groups_indexes : Vec<usize> = rev_groups.keys().map(|&x| x).collect(); 
             IndelRider{ groups: rev_groups, indexes: groups_indexes, next_group: 0, n_samples_tot: n_samples}
-            //  IndelRider{ groups: rev_groups, next_group: 0, n_samples_tot: n_samples}
         }
             
         /// Function that assigns chr samples to different groups depending on their overlapping indel alleles.
@@ -70,7 +68,7 @@
         /// * `groups` - a mutable reference to a Vec of usize that will be filled with groups info. Indexes: samples id. Elements: groups id.
         /// * `n_sample` - the number of samples (each with two alleles for each SNP) for which we have genotypes.
         ///
-        pub fn count_groups(snps_buffer: & VecDeque<mutations::Mutation>, n_overlapping: u32, groups: &mut Vec<usize>, n_samples: usize) {
+        pub fn count_groups(snps_buffer: & VecDeque<mutations::Mutation>, n_overlapping: usize, groups: &mut Vec<usize>, n_samples: usize) {
             for (i_snp, snp) in snps_buffer.iter().enumerate() {
                 if snp.is_indel && i_snp < n_overlapping as usize { // i >= n_overlapping we have finished the overlapping snps (the last one is just waiting in the buffer)
                     if snp.genotypes.iter().any(|x| x.0 || x.1) {
@@ -108,18 +106,18 @@
         }
             
         // Function that given a group id and a window will return info on the overlapping SNPs for that group and on the resulting window length
-        pub fn get_group_info(&self, window: & mut mutations::Coordinate, next_pos: & mut u64, snps_buffer: & mut VecDeque<mutations::Mutation>, n_overlapping: u32, info: & mut Vec<(usize, MutationClass)>) {
+        pub fn get_group_info(&self, window: & mut mutations::Coordinate, next_pos: & mut u64, snps_buffer: & mut VecDeque<mutations::Mutation>, n_overlapping: usize, info: & mut Vec<(usize, MutationClass)>) {
             // Right now the logic is a bit twisted cause we change coords for snps when we get a deletion but we change window.end for overlapping indels...
             // I got why I was changing in ends...to catch their overlap across window borders, but that is wrong. We need to find a way to manage indels across window borders.
             let mut pos_managed: bool = false;
             let mut last_del_start_win: bool = false;
             for (i_snp, snp) in snps_buffer.iter_mut().enumerate() {
-                if i_snp < n_overlapping as usize { // i >= n_overlapping we have finished the overlapping snps (the last one is just waiting in the buffer)
+                if i_snp < n_overlapping { // i >= n_overlapping we have finished the overlapping snps (the last one is just waiting in the buffer)
                     let next_group_i = self.indexes[self.next_group-1];
                     let this_g_v = self.groups.get(&(next_group_i)).unwrap();
                     let mut group_genotypes : Vec<bool> = Vec::with_capacity(this_g_v.len());
                     for i_sample in this_g_v.to_vec() {
-                        let index = i_sample as usize % self.n_samples_tot;
+                        let index = i_sample % self.n_samples_tot;
                         if i_sample < self.n_samples_tot {
                             group_genotypes.push(snp.genotypes.get(index).unwrap().0);
                             //println!("for sample {} seen {}", i_sample, snp.genotypes.get(index).unwrap().0);
